Update elevatorControl.cpp
After fixing the boards serial line problem the photon was able to read the serial input and I made some changes to the response string by adding a terminating char '\0'. It seemed to fix the problem of outputting extra random characters and then returning a true statement of the two strings matching.